requirements.txt
#python3.10.12

boto3==1.34.88
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
psycopg2==2.9.
python-dotenv==1.0.1
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
email_validator==2.1.1
phonenumbers==8.13.33
pillow==10.2.0
requests==2.31.0